% Generated by roxygen2 (4.0.1): do not edit by hand
\name{addListNameToReads}
\alias{addListNameToReads}
\title{Prepend name attribute of a list to DNAStringSet}
\usage{
addListNameToReads(dnaSet, flatten = FALSE)
}
\arguments{
\item{dnaSet}{output from \code{\link{extractSeqs}}}

\item{flatten}{should the output be unlisted? Default is FALSE.}
}
\value{
listed DNAStringSet with the names attribute prepended with the
name of the list. If flatten is TRUE, then a DNAStringSet object
}
\description{
Given a named listed DNAStringSet object returned from \code{\link{extractSeqs}},
the function prepends the sample name to read names.
}
\examples{
#addListNameToReads(sampleInfo)
}
\seealso{
\code{\link{extractFeature}}, \code{\link{extractSeqs}},
\code{\link{getSectorsForSamples}}, \code{\link{write.listedDNAStringSet}}
}

