% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hiReadsProcessor.R
\name{annotateSites}
\alias{annotateSites}
\title{Find the 5' primers and add results to SampleInfo object.}
\usage{
annotateSites(sampleInfo, annots = NULL, samplenames = NULL,
  parallel = TRUE, ...)
}
\arguments{
\item{sampleInfo}{sample information SimpleList object outputted from
\code{\link{findIntegrations}}, which holds genomic integration sites.}

\item{annots}{a named list of GRanges object holding features for annotating
integration sites. The name attribute of the list is used as column name.}

\item{samplenames}{a vector of samplenames to process. Default is NULL, which
processes all samples from sampleInfo object.}

\item{parallel}{use parallel backend to perform calculation. Defaults to TRUE.
If no parallel backend is registered, then a serial version is ran using
\code{\link{SerialParam}}. Parllelization is done at sample level
per sector. Use parallel2 for parallelization at sequence level.}

\item{...}{additional parameters to be passed to \code{\link{doAnnotation}}
except for sites.rd, features.rd, and colnam.}
}
\value{
a SimpleList object similar to sampleInfo paramter supplied with new
data added under each sector and sample. New data attributes include: primed
}
\description{
Given a sampleInfo object, the function finds 5' primers for each sample per
sector and adds the results back to the object. This is a specialized
function which depends on many other functions shown in 'see also section'
to perform specialized trimming of 5' primer/adaptor found in the sampleInfo
object. The sequence itself is never trimmed but rather coordinates of primer
portion is recorded back to the object and used subsequently by
\code{\link{extractSeqs}} function to perform the trimming.
}
\examples{
\donttest{
load(file.path(system.file("data", package = "hiReadsProcessor"),
"FLX_seqProps.RData"))
data(genes)
genes <- makeGRanges(genes)
cpgs <- getUCSCtable("cpgIslandExt","CpG Islands")
cpgs <- makeGRanges(cbind(cpgs,strand="*"), chromCol = "chrom")
annots <- list("RefGenes"=genes,"CpG"=cpgs)
annotateSites(seqProps, annots, annotType="nearest", side="5p")
}
}
\seealso{
\code{\link{clusterSites}}, \code{\link{isuSites}},
\code{\link{crossOverCheck}}, \code{\link{findIntegrations}},
\code{\link{getIntegrationSites}}, \code{\link{pslToRangedObject}}
}

