% Generated by roxygen2 (4.0.1): do not edit by hand
\name{doRCtest}
\alias{doRCtest}
\title{Test if pattern aligns better in +/- orientation.}
\usage{
doRCtest(subjectSeqs = NULL, patternSeq = NULL, qualityThreshold = 0.5,
  core.use = 2)
}
\arguments{
\item{subjectSeqs}{DNAStringSet object containing sequences to be searched for the pattern.}

\item{patternSeq}{DNAString object or a sequence containing the query sequence to search.}

\item{qualityThreshold}{percent of patternSeq to match. Default is 0.50, half match. This is supplied to max.mismatch parameter of \code{\link{vcountPattern}} as round(nchar(patternSeq)*(1-qualityThreshold)).}

\item{cores.use}{Number of cores to use for aligning. Only need two max! Default is 2. If no parallel backend is registered, then a serial version is ran using \code{\link{SerialParam}}.}
}
\value{
patternSeq that aligned the best.
}
\description{
Given a fixed length pattern sequence and variable length subject sequences, the function roughly finds which orientation of pattern yields the most hits. The function doing the heavylifting is \code{\link{vcountPattern}}. This is an accessory function used in function listed under See Also section below.
}
\examples{
#doRCtest(subjectSeqs,patternSeq)
}
\seealso{
\code{\link{pairwiseAlignSeqs}}, \code{\link{vpairwiseAlignSeqs}}, \code{\link{primerIDAlignSeqs}}
}

