% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiReadsProcessor.R
\name{findAndTrimSeq}
\alias{findAndTrimSeq}
\title{Find and trim a short pattern sequence from the subject.}
\usage{
findAndTrimSeq(patternSeq, subjectSeqs, side = "left", offBy = 0,
  alignWay = "slow", ...)
}
\arguments{
\item{patternSeq}{DNAString object or a sequence containing the query sequence to search.}

\item{subjectSeqs}{DNAStringSet object containing sequences to be searched for the pattern.}

\item{side}{which side of the sequence to perform the search & trimming: left, right or middle. Default is 'left'.}

\item{offBy}{integer value dictating if the trimming base should be offset by X number of bases. Default is 0.}

\item{alignWay}{method to utilize for detecting the primers. One of following: "slow" (Default), "fast", or "blat". Fast, calls \code{\link{vpairwiseAlignSeqs}} and uses \code{\link{vmatchPattern}} at its core, which is less accurate with indels and mismatches but much faster. Slow, calls \code{\link{pairwiseAlignSeqs}} and uses \code{\link{pairwiseAlignment}} at its core, which is accurate with indels and mismatches but slower. Blat will use \code{\link{blatSeqs}}.}

\item{...}{parameters to be passed to \code{\link{pairwiseAlignment}}, \code{\link{vpairwiseAlignSeqs}} or \code{\link{blatSeqs}} depending on which method is defined in 'alignWay' parameter.}
}
\value{
DNAStringSet object with pattern sequence removed from the subject sequences.
}
\description{
This function facilitates finding and trimming of a short pattern sequence from a collection of subject sequences. The trimming is dictated by side parameter. For more information on the trimming process see the 'side' parameter documentation in \code{\link{trimSeqs}}. For information regarding the pattern alignment see the documentation for \code{\link{pairwiseAlignSeqs}}. This function is meant for aligning a short pattern onto large collection of subjects. If you are looking to align a vector sequence to subjects, then please use BLAT.
}
\note{
If parallel=TRUE, then be sure to have a parallel backend registered 
before running the function. One can use any of the following 
\code{\link{MulticoreParam}} \code{\link{SnowParam}}
}
\examples{
findAndTrimSeq(patternSeq="AGACCCTTTT",
subjectSeqs=DNAStringSet(c("AGACCCTTTTGAGCAGCAT","AGACCCTTGGTCGACTCA",
"AGACCCTTTTGACGAGCTAG")), qualityThreshold=.85, doRC=FALSE, side="left", 
offBy=1, alignWay = "slow")

}
\seealso{
\code{\link{pairwiseAlignSeqs}}, \code{\link{vpairwiseAlignSeqs}}, \code{\link{extractFeature}}, \code{\link{extractSeqs}}, \code{\link{primerIDAlignSeqs}}, \code{\link{findPrimers}}, \code{\link{findLinkers}}
}

