% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiReadsProcessor.R
\name{findPrimers}
\alias{findPrimers}
\title{Find the 5' primers and add results to SampleInfo object.}
\usage{
findPrimers(sampleInfo, alignWay = "slow", showStats = FALSE,
  doRC = FALSE, parallel = TRUE, samplenames = NULL,
  bypassChecks = FALSE, parallel2 = FALSE, ...)
}
\arguments{
\item{sampleInfo}{sample information SimpleList object outputted from 
\code{\link{findBarcodes}}, which holds decoded sequences for samples 
per sector/quadrant along with information of sample to primer associations.}

\item{alignWay}{method to utilize for detecting the primers. One of 
following: "slow" (Default), or "fast". Fast, calls 
\code{\link{vpairwiseAlignSeqs}} and uses \code{\link{vmatchPattern}} at its 
core, which is less accurate with indels and mismatches but much faster. 
Slow, calls \code{\link{pairwiseAlignSeqs}} and uses 
\code{\link{pairwiseAlignment}} at its core, which is accurate with indels 
and mismatches but slower.}

\item{showStats}{toggle output of search statistics. Default is FALSE.}

\item{doRC}{perform reverse complement search of the defined pattern/primer. 
Default is FALSE.}

\item{parallel}{use parallel backend to perform calculation . Defaults to TRUE. 
If no parallel backend is registered, then a serial version is ran using
\code{\link{SerialParam}}. Parllelization is done at sample level 
per sector. Use parallel2 for parallelization at sequence level.}

\item{samplenames}{a vector of samplenames to process. Default is NULL, which
processes all samples from sampleInfo object.}

\item{bypassChecks}{skip checkpoints which detect if something was odd with 
the data? Default is FALSE.}

\item{parallel2}{perform parallelization is sequence level. Default is FALSE.
Useful in cases where each sector has only one sample with numerous sequences.}

\item{...}{extra parameters to be passed to either \code{\link{vmatchPattern}} 
or \code{\link{pairwiseAlignment}} depending on 'alignWay' parameter.}
}
\value{
a SimpleList object similar to sampleInfo paramter supplied with new 
data added under each sector and sample. New data attributes include: primed
}
\description{
Given a sampleInfo object, the function finds 5' primers for each sample per 
sector and adds the results back to the object. This is a specialized 
function which depends on many other functions shown in 'see also section' 
to perform specialized trimming of 5' primer/adaptor found in the sampleInfo 
object. The sequence itself is never trimmed but rather coordinates of primer
portion is recorded back to the object and used subsequently by 
\code{\link{extractSeqs}} function to perform the trimming.
}
\note{
\itemize{
 \item For paired end data, qualityThreshold for pair 2 is decreased by 
 0.10 to increase chances of matching primer sequence.
 \item If parallel=TRUE, then be sure to have a parallel backend registered 
 before running the function. One can use any of the following 
 \code{\link{MulticoreParam}} \code{\link{SnowParam}}
}
}
\examples{

\donttest{
load(file.path(system.file("data", package = "hiReadsProcessor"),
"FLX_seqProps.RData"))
findPrimers(seqProps, showStats=TRUE)
}
}
\seealso{
\code{\link{pairwiseAlignSeqs}}, \code{\link{vpairwiseAlignSeqs}},
\code{\link{extractFeature}}, \code{\link{extractSeqs}}, 
\code{\link{primerIDAlignSeqs}}, \code{\link{findLTRs}}, 
\code{\link{findLinkers}}, \code{\link{findAndTrimSeq}}
}

