% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiReadsProcessor.R
\name{findVector}
\alias{findVector}
\title{Find vector DNA in reads and add results to SampleInfo object.}
\usage{
findVector(sampleInfo, showStats = FALSE, parallel = TRUE, samplenames = NULL)
}
\arguments{
\item{sampleInfo}{sample information SimpleList object outputted from 
\code{\link{findLTRs}}, which holds decoded, primed, and LTRed sequences for 
samples per sector/quadrant.}

\item{showStats}{toggle output of search statistics. Default is FALSE.}

\item{parallel}{use parallel backend to perform calculation with 
\code{\link{BiocParallel}}. Defaults to TRUE. If no parallel backend is  
registered, then a serial version is ran using \code{\link{SerialParam}}.
Parllelization is done at sample level per sector. Use parallel2 for 
parallelization at sequence level.}

\item{samplenames}{a vector of samplenames to process. Default is NULL, which
processes all samples from sampleInfo object.}
}
\value{
a SimpleList object similar to sampleInfo paramter supplied with new 
data added under each sector and sample. New data attributes include: vectored
}
\description{
Given a sampleInfo object, the function finds vector fragments following the 
LTR piece for each sample per sector and adds the results back to the object. 
This is a specialized function which depends on many other functions shown
in 'see also section' to perform specialized trimming of 5' viral LTRs found
in the sampleInfo object. The sequence itself is never trimmed but rather 
coordinates of vector portion is added to LTR coordinates and recorded back 
to the object and used subsequently by \code{\link{extractSeqs}} function to 
perform the trimming. This function heavily relies on \code{\link{blatSeqs}}.
In order for this function to work, it needs vector sequence which is read in
using 'vectorFile' metadata supplied in the sample information file in 
\code{\link{read.sampleInfo}}
}
\note{
\itemize{
 \item If parallel=TRUE, then be sure to have a parallel backend registered 
 before running the function. One can use any of the following 
 \code{\link{MulticoreParam}} \code{\link{SnowParam}}
}
}
\examples{

\dontrun{
load(file.path(system.file("data", package = "hiReadsProcessor"),
"FLX_seqProps.RData"))
findVector(seqProps, showStats=TRUE)
}
}
\seealso{
\code{\link{pairwiseAlignSeqs}}, \code{\link{blatSeqs}},
\code{\link{extractFeature}}, \code{\link{extractSeqs}}, 
\code{\link{findPrimers}}, \code{\link{findLTRs}}, 
\code{\link{findLinkers}}, \code{\link{findAndTrimSeq}},
\code{\link{findAndRemoveVector}}
}
