% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getIntegrationSites}
\alias{getIntegrationSites}
\title{Obtain integration sites from BLAT output}
\usage{
getIntegrationSites(psl.rd = NULL, startWithin = 3,
  alignRatioThreshold = 0.7, genomicPercentIdentity = 0.98,
  correctByqStart = TRUE, oneBased = FALSE)
}
\arguments{
\item{psl.rd}{a GRanges object reflecting psl format where tName is
the seqnames.}

\item{startWithin}{upper bound limit of where the alignment should start
within the query. Default is 3.}

\item{alignRatioThreshold}{cuttoff for (alignment span/read length).
Default is 0.7.}

\item{genomicPercentIdentity}{cuttoff for (1-(misMatches/matches)).
Default is 0.98.}

\item{correctByqStart}{use qStart to correct genomic position. This would
account for sequencing/trimming errors.
Position=ifelse(strand=="+",tStart-qStart,tEnd+qStart). Default is TRUE.}

\item{oneBased}{the coordinates in psl files are "zero based half open".
The first base in a sequence is numbered zero rather than one. Enabling
this would add +1 to the start and leave the end as is. Default is FALSE.}
}
\value{
a GRanges object with integration sites which passed all filtering
criteria. Each filtering parameter creates a new column to flag if a
sequence/read passed that filter which follows the scheme:
'pass.FilterName'. Integration Site is marked by new column named 'Position'.
}
\description{
Given a GRanges object from \code{\link{read.psl}}, the function uses
specified filtering parameters to obtain integration sites and maintain
sequence attrition. The function will remove any non-best scoring alignments
from the object if not already filtered apriori.
}
\examples{
data(psl)
psl.rd <- pslToRangedObject(psl)
getIntegrationSites(psl.rd)
}
\seealso{
\code{\link{startgfServer}}, \code{\link{read.psl}},
\code{\link{blatSeqs}}, \code{\link{blatListedSet}},
\code{\link{findIntegrations}}, \code{\link{pslToRangedObject}},
\code{\link{clusterSites}}, \code{\link{isuSites}},
\code{\link{crossOverCheck}}, \code{\link{read.blast8}}
}

