\docType{package}
\name{hiReadsProcessor}
\alias{hiReadsProcessor}
\alias{hiReadsProcessor-package}
\title{Functions to process LM-PCR reads from 454/Illumina data.}
\description{
hiReadsProcessor contains set of functions which allow
users to process LM-PCR products sequenced using any
platform. Given an excel/txt file containing parameters for
demultiplexing and sample metadata, the functions automate
trimming of adaptors andidentification of the genomic
product. In addition, if IntSites MySQL database is setup,
the sequence attrition is loaded into respective tables for
post processing setup and analysis.
}
\author{
Nirav V Malani
}

