% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiReadsProcessor.R
\name{pslCols}
\alias{pslCols}
\title{Return PSL file columns with classes}
\usage{
pslCols(withClass = TRUE)
}
\arguments{
\item{withClass}{return classes for each column.}
}
\value{
vector of PSL column names
}
\description{
Print out required fields & classes of PSL file format
}
\examples{
pslCols()

}
\seealso{
\code{\link{pairwiseAlignSeqs}}, \code{\link{vpairwiseAlignSeqs}}, 
\code{\link{startgfServer}}, \code{\link{blatSeqs}}, \code{\link{read.blast8}}, 
\code{\link{read.BAMasPSL}}, \code{\link{pslToRangedObject}}
}

