% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiReadsProcessor.R
\name{pslToRangedObject}
\alias{pslToRangedObject}
\title{Convert psl dataframe to GRanges}
\usage{
pslToRangedObject(x, useTargetAsRef = TRUE, isblast8 = FALSE)
}
\arguments{
\item{x}{dataframe reflecting psl format}

\item{useTargetAsRef}{use target(tName) or query(qName) as the chromosome or 
the reference data. Default is TRUE.}

\item{isblast8}{the input dataframe blast8 format output from BLAT. 
Default is FALSE.}
}
\value{
a GRanges object reflecting psl file type.
}
\description{
Convert psl dataframe to GRanges object using either the query or target as 
the reference data column.
}
\examples{
data(psl)
psl <- head(psl)
pslToRangedObject(psl)
pslToRangedObject(psl, useTargetAsRef = FALSE)
}
\seealso{
\code{\link{read.psl}}, \code{\link{read.blast8}}, 
\code{\link{blatListedSet}}
}

