% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiReadsProcessor.R
\name{read.blast8}
\alias{read.blast8}
\title{Read blast8 file(s) outputted by BLAT}
\usage{
read.blast8(
  files = NULL,
  asGRanges = FALSE,
  removeFile = TRUE,
  parallel = FALSE
)
}
\arguments{
\item{files}{blast8 filename, or vector of filenames, or a pattern of files 
to import.}

\item{asGRanges}{return a GRanges object instead of a dataframe.
Default is TRUE Saves memory!}

\item{removeFile}{remove the blast8 file(s) after importing. 
Default is FALSE.}

\item{parallel}{use parallel backend to perform calculation with 
\code{\link{BiocParallel}}. Defaults to TRUE. If no parallel backend is 
registered, then a serial version is ran using \code{\link{SerialParam}}.}
}
\value{
a dataframe or GRanges object reflecting blast8 file type.
}
\description{
Given filename(s), the function reads the blast8 file format from BLAT as a 
data frame and performs basic score filtering if indicated. Any other file 
format will yield errors or erroneous results.
}
\note{
If parallel=TRUE, then be sure to have a parallel backend registered 
before running the function. One can use any of the following 
\code{\link{MulticoreParam}} \code{\link{SnowParam}}
}
\examples{
# this function works similar to read.psl #
#read.blast8(files="processed.*.blast8$")
#read.blast8(files=c("sample1hits.blast8","sample2hits.blast8"))

}
\seealso{
\code{\link{pairwiseAlignSeqs}}, \code{\link{vpairwiseAlignSeqs}},
\code{\link{startgfServer}}, \code{\link{blatSeqs}}, \code{\link{read.psl}}
}
