% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiReadsProcessor.R
\name{replicateReads}
\alias{replicateReads}
\title{Replicate sequences from DNAStringSet object using counts identifier or vector}
\usage{
replicateReads(dnaSet, counts = NULL)
}
\arguments{
\item{dnaSet}{DNAStringSet object to replicate.}

\item{counts}{an integer or a numeric vector of length length(dnaSet) 
indicating how many times to repeat each sequence. Default is NULL, 
in which it uses counts=X notation from the definition line to 
replicate reads.}
}
\value{
DNAStringSet object.
}
\description{
Given a DNAStringSet object, the function replicates reads using counts=X 
marker at the end of definition line.
}
\examples{
 
dnaSet <- c("CCTGAATCCTGGCAATGTCATCATC", "ATCCTGGCAATGTCATCATCAATGG", 
"ATCAGTTGTCAACGGCTAATACGCG", "ATCAATGGCGATTGCCGCGTCTGCA", 
"CCGCGTCTGCAATGTGAGGGCCTAA", "GAAGGATGCCAGTTGAAGTTCACAC", 
"CCTGAATCCTGGCAATGTCATCATC", "ATCCTGGCAATGTCATCATCAATGG", 
"ATCAGTTGTCAACGGCTAATACGCG", "ATCAATGGCGATTGCCGCGTCTGCA", 
"CCGCGTCTGCAATGTGAGGGCCTAA", "GAAGGATGCCAGTTGAAGTTCACAC") 
dnaSet <- dereplicateReads(dnaSet)
replicateReads(dnaSet)
}
\seealso{
\code{\link{dereplicateReads}}, \code{\link{removeReadsWithNs}},
\code{\link{findBarcodes}}, \code{\link{splitByBarcode}}
}
