% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiReadsProcessor.R
\docType{data}
\name{seqProps}
\alias{seqProps}
\title{Sample Integration Sites Sequencing Data}
\format{A SimpleList object}
\description{
This is a processed data object containing raw sequences and respective 
alignments to UCSC freeze hg18 from 112 integration site samples. The object
is of SimpleList class and follows a certain structural hierarchy explained
by the Introductory vignette.
}
\keyword{datasets}
