% Generated by roxygen2 (4.0.1): do not edit by hand
\name{splitSeqsToFiles}
\alias{splitSeqsToFiles}
\title{Split DNA sequences into smaller files.}
\usage{
splitSeqsToFiles(x, totalFiles = 4, suffix = "tempy",
  filename = "queryFile.fa")
}
\arguments{
\item{x}{a DNAStringSet object, or a FASTA filename.}

\item{totalFiles}{an integer indicating how many files to create. Default is 4.}

\item{suffix}{a word to add to each file created. Default is "tempy".}

\item{filename}{name of the file if x is a DNAStringSet object.
Default is "queryFile.fa".}
}
\value{
a vector of filename names created.
}
\description{
Given a vector of sequences or DNAStringSet or a FASTA filename, the function
splits it into smaller pieces as denoted by totalFiles parameter.
}
\examples{
\donttest{
seqs <- DNAStringSet(sapply(sample(c(100:1000), 500),
function(size) paste(sample(DNA_BASES, size, replace=TRUE), collapse="")))
splitSeqsToFiles(seqs,5,"tempyQ","myDNAseqs.fa")
}
}
\seealso{
\code{\link{blatSeqs}}
}

