\name{summary.elegant}
\alias{summary.elegant}
\title{Elegant summary of a sampleInfo object.
Give an elegant summary of all the attributes in sampleInfo/SimpleList object.}
\usage{
  summary.elegant(sampleInfo, samplenames = NULL)
}
\arguments{
  \item{sampleInfo}{sample information SimpleList object,
  which samples per sector/quadrant information along with
  other metadata.}

  \item{samplenames}{a vector of samplenames to process.
  Default is NULL, which processes all samples from
  sampleInfo object.}
}
\description{
  Elegant summary of a sampleInfo object. Give an elegant
  summary of all the attributes in sampleInfo/SimpleList
  object.
}
\examples{
\dontrun{
 summary.elegant(sampleInfo)
}
}

