\name{summary.simple}
\alias{summary.simple}
\title{Simple summary of a sampleInfo object.
Give a simple summary of major attributes in sampleInfo/SimpleList object.}
\usage{
  summary.simple(sampleInfo)
}
\arguments{
  \item{sampleInfo}{sample information SimpleList object,
  which samples per sector/quadrant information along with
  other metadata.}
}
\description{
  Simple summary of a sampleInfo object. Give a simple
  summary of major attributes in sampleInfo/SimpleList
  object.
}
\examples{
\dontrun{
 summary.simple(sampleInfo)
}
}

