% Generated by roxygen2 (4.0.0): do not edit by hand
\name{trimSeqs}
\alias{trimSeqs}
\title{Trim sequences from a specific side.}
\usage{
trimSeqs(dnaSet, coords, side = "middle", offBy = 0)
}
\arguments{
\item{dnaSet}{DNAStringSet object containing sequences to be trimmed.}

\item{coords}{IRanges object containing boundaries.}

\item{side}{either 'left','right',or the Default 'middle'.}

\item{offBy}{integer value dictating if the supplied coordinates should be offset by X number of bases. Default is 0.}
}
\value{
a DNAStringSet object with trimmed sequences.
}
\description{
This function trims a DNAStringSet object using the ranges from left, right, or middle of the sequence. This is a helper function utilized in \code{\link{primerIDAlignSeqs}} and \code{\link{extractSeqs}}. If dnaSet and coords are not the same length, then they are required to have a names attribute to perform the matched trimming.
}
\note{
If side is left, then any sequence following end of coords+offBy is returned. If side is right, then sequence preceding start of coords-offBy is returned. If side is middle, then sequence contained in coords is returned where offBy is added to start and subtracted from end in coords.
}
\examples{
#trimSeqs(dnaSet,coords)
#trimSeqs(dnaSet,coords,side="left",offBy=1)
}
\seealso{
\code{\link{extractSeqs}}, \code{\link{primerIDAlignSeqs}}
}

