% Generated by roxygen2 (4.0.1): do not edit by hand
\name{write.runData}
\alias{write.runData}
\title{Write the information from sample information object back to disk.}
\usage{
write.runData(sampleInfo, outPath = NULL)
}
\arguments{
\item{sampleInfo}{sample information SimpleList object which has been through
at least \code{\link{findBarcodes}}, \code{\link{findPrimers}},
\code{\link{findLTRs}}, \code{\link{findLinkers}},
\code{\link{findIntegrations}}}

\item{outPath}{path to output location. Defaults to 'SequencingFolderPath'
attribute of the sampleInfo object.}
}
\value{
location of the exported run data
}
\description{
Given a sampleInfo object, the function will reconstruct the sequencing run
folder with sector/lane files and relevant sampleInformation file. If
genomic alignment attributes were recorded, then each sample sites will be
outputted into a text file in the root of run folder.
}
\note{
if the run folder already exists, then the data will be overwritten!
}
\examples{
#write.runData(sampleInfo)
}
\seealso{
\code{\link{extractFeature}}, \code{\link{extractSeqs}},
\code{\link{getSectorsForSamples}}
}

