% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiReadsProcessor.R
\name{addListNameToReads}
\alias{addListNameToReads}
\title{Prepend name attribute of a list to DNAStringSet}
\usage{
addListNameToReads(dnaSet, flatten = FALSE)
}
\arguments{
\item{dnaSet}{output from \code{\link{extractSeqs}}}

\item{flatten}{should the output be unlisted? Default is FALSE.}
}
\value{
listed DNAStringSet with the names attribute prepended with the 
name of the list. If flatten is TRUE, then a DNAStringSet object
}
\description{
Given a named listed DNAStringSet object returned from \code{\link{extractSeqs}}, 
the function prepends the sample name to read names.
}
\examples{
load(file.path(system.file("data", package = "hiReadsProcessor"),
"FLX_seqProps.RData"))
samples <- c('Roth-MLV3p-CD4TMLVWell6-Tsp509I', 
'Roth-MLV3p-CD4TMLVWell6-MseI', 'Roth-MLV3p-CD4TMLVwell5-MuA')
seqs <- extractSeqs(seqProps, sector = '2', samplename = samples,
feature="genomic")
addListNameToReads(seqs, TRUE)
}
\seealso{
\code{\link{extractFeature}}, \code{\link{extractSeqs}}, 
\code{\link{getSectorsForSamples}}, \code{\link{write.listedDNAStringSet}}
}
