% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiReadsProcessor.R
\name{chunkize}
\alias{chunkize}
\title{Breaks an object into chunks of N size.}
\usage{
chunkize(x, chunkSize = NULL)
}
\arguments{
\item{x}{a linear object.}

\item{chunkSize}{number of rows to use per chunk of query. 
Defaults to length(x)/detectCores() or length(query)/bpworkers() 
depending on parallel backend registered.}
}
\value{
a list of object split into chunks.
}
\description{
Given a linear/vector like object, the function breaks it into equal sized chunks either by chunkSize. This is a helper function used by functions in 'See Also' section where each chunk is sent to a parallel node for processing.
}
\examples{
x <- c("GAGGCTGTCACCGACAAGGTTCTGA", "AATAGCGTGGTGACAGCCCACATGC", 
       "GGTCTTCTAGGGAACCTACGCCACA", "TTTCCGGCGGCAGTCAGAGCCAAAG", 
       "TCCTGTCAACTCGTAGATCCAATCA", "ATGGTCACCTACACACAACGGCAAT", 
       "GTCAGGACACCTAATCACAAACGGA", "AGACGCAGGTTCAGGCTGGACTAGA", 
       "ATCGTTTCCGGAATTCGTGCACTGG", "CAATGCGGGCACACGCTCTTACAGT")
chunkize(DNAStringSet(x), 5)
}
\seealso{
\code{\link{primerIDAlignSeqs}}, \code{\link{vpairwiseAlignSeqs}}, 
\code{\link{pairwiseAlignSeqs}}
}
