% Generated by roxygen2 (4.0.1): do not edit by hand
\name{findLinkers}
\alias{findLinkers}
\title{Find the 3' linkers and add results to SampleInfo object.}
\usage{
findLinkers(sampleInfo, showStats = FALSE, doRC = FALSE, parallel = TRUE,
  samplenames = NULL, bypassChecks = FALSE, parallel2 = FALSE, ...)
}
\arguments{
\item{sampleInfo}{sample information SimpleList object outputted from
\code{\link{findPrimers}} or \code{\link{findLTRs}}, which holds decoded
sequences for samples per sector/quadrant along with information of sample
to primer associations.}

\item{showStats}{toggle output of search statistics. Default is FALSE.}

\item{doRC}{perform reverse complement search of the defined pattern/linker
sequence. Default is FALSE.}

\item{parallel}{use parallel backend to perform calculation with
\code{\link{BiocParallel}}. Defaults to TRUE. If no parallel backend is
registered, then a serial version is ran using
\code{\link{SerialParam}}. Parllelization is done at sample level
per sector.}

\item{samplenames}{a vector of samplenames to process. Default is NULL,
which processes all samples from sampleInfo object.}

\item{bypassChecks}{skip checkpoints which detect if something was odd
with the data? Default is FALSE.}

\item{parallel2}{perform parallelization is sequence level. Default is FALSE.
Useful in cases where each sector has only one sample with numerous sequences.}

\item{...}{extra parameters to be passed to \code{\link{pairwiseAlignment}}.}
}
\value{
a SimpleList object similar to sampleInfo paramter supplied with new
data added under each sector and sample. New data attributes include:
linkered. If linkers have primerID then, primerIDs attribute is appended
as well.
}
\description{
Given a sampleInfo object, the function finds 3' linkers for each sample per
sector and adds the results back to the object. This is a specialized
function which depends on many other functions shown in 'see also section' to
perform specialized trimming of 3' primer/linker adaptor sequence found in
the sampleInfo object. The sequence itself is never trimmed but rather
coordinates of linker portion is recorded back to the object and used
subsequently by \code{\link{extractSeqs}} function to perform the trimming.
This function heavily relies on either \code{\link{pairwiseAlignSeqs}} or
\code{\link{primerIDAlignSeqs}} depending upon whether linkers getting
aligned have primerID in it or not.
}
\note{
\itemize{
 \item For paired end data, qualityThreshold for pair 2 is increased by
 0.25 or set to 1 whichever is lower to increase quality & full match to
 linker sequence.
 \item If no linker matches are found with default options, then try
 doRC=TRUE.
 \item If parallel=TRUE, then be sure to have a parallel backend registered
 before running the function. One can use any of the following
 \code{\link{MulticoreParam}} \code{\link{SnowParam}}
}
}
\examples{
\donttest{
load(file.path(system.file("data", package = "hiReadsProcessor"),
"FLX_seqProps.RData"))
findLinkers(seqProps, showStats=TRUE, doRC=TRUE)
}
}
\seealso{
\code{\link{pairwiseAlignSeqs}}, \code{\link{vpairwiseAlignSeqs}},
\code{\link{primerIDAlignSeqs}}, \code{\link{findLTRs}},
\code{\link{findPrimers}}, \code{\link{extractFeature}},
\code{\link{extractSeqs}}, \code{\link{findAndTrimSeq}},
\code{\link{findIntegrations}}
}

