% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiReadsProcessor.R
\docType{data}
\name{psl}
\alias{psl}
\title{PSL file output}
\format{a data frame of 1000 rows and 21 columns}
\description{
Sample BLAT PSL file output from samples included Integration Sites 
Sequencing Data \code{\link{seqProps}}
}
\keyword{datasets}

