% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiReadsProcessor.R
\name{read.seqsFromSector}
\alias{read.seqsFromSector}
\title{Read fasta/fastq/sff given the path or sampleInfo object.}
\usage{
read.seqsFromSector(seqFilePath = NULL, sector = 1, isPaired = FALSE)
}
\arguments{
\item{seqFilePath}{a path to fasta/fastq/sff reads file or a sampleInfo 
object returned by \code{\link{read.SeqFolder}}}

\item{sector}{specific sector to reads sequences from. Default is 1, and not 
required if seqFilePath is a direct file path rather than sampleInfo object.}

\item{isPaired}{does the sector contain paired end reads? Default is FALSE}
}
\value{
if isPaired is FALSE, then a DNAStringSet object, else a list of 
DNAStringSet objects of three elements corresponding to reads from 
"barcode", "pair1", and "pair2". Note: "pair2" is reverse complemented!
}
\description{
Given a sequence reads file path, the function returns a DNAStringSet object.
}
\examples{
\donttest{
load(file.path(system.file("data", package = "hiReadsProcessor"),
"FLX_seqProps.RData"))
read.seqsFromSector(seqProps, sector="2")
}
}
\seealso{
\code{\link{findBarcodes}}, \code{\link{read.SeqFolder}}, 
\code{\link{extractSeqs}}
}

