% Generated by roxygen2 (4.0.0): do not edit by hand
\name{replicateReads}
\alias{replicateReads}
\title{Replicate sequences from DNAStringSet object using counts identifier or vector}
\usage{
replicateReads(dnaSet, counts = NULL)
}
\arguments{
\item{dnaSet}{DNAStringSet object to replicate.}

\item{counts}{an integer or a numeric vector of length length(dnaSet) indicating how many times to repeat each sequence. Default is NULL, in which it uses counts=X notation from the definition line to replicate reads.}
}
\value{
DNAStringSet object.
}
\description{
Given a DNAStringSet object, the function replicates reads using counts=X
marker at the end of definition line.
}
\examples{
#replicateReads(decoded)
}
\seealso{
\code{\link{dereplicateReads}}, \code{\link{removeReadsWithNs}},
\code{\link{decodeByBarcode}}, \code{\link{splitByBarcode}}
}

