% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiReadsProcessor.R
\name{sampleSummary}
\alias{sampleSummary}
\title{Simple summary of a sampleInfo object.}
\usage{
sampleSummary(object, ...)
}
\arguments{
\item{object}{sample information SimpleList object, which samples per 
sector/quadrant information along with other metadata.}

\item{...}{ignored for now.}
}
\value{
a dataframe summarizing counts of major attributes per sample 
and sector.
}
\description{
Give a simple summary of major attributes in sampleInfo/SimpleList object.
}
\examples{
data(FLX_seqProps)
sampleSummary(seqProps)

}
