% Generated by roxygen2 (4.0.1): do not edit by hand
\name{stopgfServer}
\alias{stopgfServer}
\title{Stop a gfServer instance}
\usage{
stopgfServer(host = "localhost", port = NULL)
}
\arguments{
\item{host}{name of the machine running gfServer. Default: localhost}

\item{port}{the same number you started the gfServer with.}
}
\value{
system command status for executing kill command.
}
\description{
Take down the gfServer instance.
}
\examples{
#stopgfServer(port=5560)
}
\seealso{
\code{\link{startgfServer}}
}

