\name{summary.simple}
\alias{summary.simple}
\title{Simple summary of a sampleInfo object.}
\usage{
summary.simple(sampleInfo)
}
\arguments{
  \item{sampleInfo}{sample information SimpleList object,
  which samples per sector/quadrant information along with
  other metadata.}
}
\value{
a dataframe summarizing counts of major attributes per
sample and sector.
}
\description{
Give a simple summary of major attributes in
sampleInfo/SimpleList object.
}
\examples{
#summary.simple(sampleInfo)
}

